Added validation test cases and also the Xunit test for services and controllers

+++ localink_be/Services/Implementations/ContactService.cs
@@ -55,7 +55,7 @@ public class ContactService : IContactService
         await _db.SaveChangesAsync();
         return existing;
     }
-    public async Task<bool> DeleteContactAsync(int contactId)
+    public async Task<bool> DeleteContactAsync(long contactId)
     {
         var contact = await _db.BusinessContacts.FindAsync(contactId);
         if (contact == null) return false;
